Big RAD commit: enhancements, bug fixes and tweaks following extensive testing, Documentation upgrade.
configure.ac
	- add -I /usr/src/sys  to CFLAGS so kernel versions of bpf.h, timeffc.h used

Makefile.am
	- includes tests in DIST_SUBDIRS to prevent strange error in Makefiles
	- adding additional dirs to include:  FreeBSD, Tools
	- commenting out unneeded parts of dist-hook
	- adding new entries for new test programs in Tests/

Library
   - fixed bug in read_RADabs_{native,UTC} where readings using plocal would be
     potentially incorrect if the vcount was in the past wrt the rad_data used.
	- replacing (1LLU << 63) followed by <<1 logic with simpler alternatives
	  with no precision loss by defining  #define TWO32
	- defining  bintime_to_ld  and  ld_to_bintime  functions, using in testing
		and fill_ffclock_estimate

MAIN thread
 rehash_daemon()
 	- bug fix where NTP_SERV restarted, when already done in start_live
 	- bug fix failing to reinitialize verbosity level following change
 	- improved verbosity
 start_live()
 	- verbosity correction and comment improvements
 livepcapstamp_update_filter()
 	- new hostIP member added to handle so it can be reused here during restarts
 config_parse()
	- fixed bug in logfile name setting

PROC thread
 bad_packet_server
 	- Detection of NTP_VERSION=5 as work-in-progress hack to label a server as
 	  a RADclock server to clients
 create_stamp.c
 	- two bug fixes in refid verbosity
 	- enhancement of verbosity detailing detected changes in server status
 pthread_dataproc.c
 	- improved ordering, classification and verbosity of insane stamp detection
 	- addition of adjust_FFclock  configuration parameter, to allow the pushing
 	  of updates to the FFclock to be blocked if desired.
	- detection of RTC event from the kernel, with reaction being a warning
	  message and daemon termination:  cannot currently deal with an interrupted
	  counter safely, risk of pushing erroneous updates to FFclock.
	  A restart is safe.
 pthread_mgr.c
 	- correction and improvement of comments in thread_{data_processing,trigger}
 	  describing how the signaling works between these tag-team threads
 radclock.h
 	- new tsmode  PKTCAP_TSMODE_CUSTOM = 100	allowing a custom tsmode to be
 	  selected  (not needed for daemon, used by test programs)
 syncalgo.h: reinstatement of old value  SKM_SCALE_GOOD = 1024 !

 TRIGGER thread
 create_ntp_request.c:
 	- correction of refid setting in NTP request
 ntp_client_init.c:
 	- adjustment of #define timeout parameters to more general values
	- suppression of retries if poll_period <= 2
	- adding `second look` to retries to solve problem of out-of-sync `matching`
	- improved timeout setting code as a fn of RTT, now safer, more efficient
	- more extensive comments and improved verbosity, including adding
	  NTPclient prefix to non-DEBUG messages

NTP_SERV thread
 thread_ntp_server.c:
 	- improved organisation to aid readability
 	- correction of LI code in case of stratum=0
 	- correction of refid setting in NTP response [ see comment block ]
	- use of read_RADabs_UTC and UTCld_to_NTPtime  in timestamp generation,
	  mirroring that of TRIGGER side
	- use of Difference clock to generate xmt timestamp (more robust, accurate)
	- bug fix: socket wasnt closed when thread died
	- more extensive comments and improved verbosity, including adding
	  NTPserver prefix to non-DEBUG messages

Testing
	- enhanced  examples/radclock_capture_packet.c   showing how to play with
	  customized tsmodes
	- new tests/test_FFclock.c  for in-depth testing of the FFclocks and tsmode
	  options, and comparison of kernel and radclock timestamps off same packets.
	- new tests/test_clockcompare.c  for in-depth comparison of all kernel clocks
	  (FF and FB), using shared raw timestamp ability in bpf, triggered off
	  packets matching provided filter.

Tools  [ New directory to hold convenience scripts ]
	- SHMset, confset_{binary,value}  to streamline reliable conf file changes.
	- install_startscript	:  wrapper to install OS-specific startup scripts
	- update_FFkernelsource :  wrapper to run OS-specific scripts to
	  move all needed FFclock sources files into the source tree.  This is a
	  better alternative to the patch also provided as they work on top of source
	  which may have been updated from standard releases.
	- compile_FFkernel		: script to build and install the FF kernel

Creation of  FreeBSD directory for FreeBSD specific installation scripts
	- FreeBSD daemon_startup_script :  script for rd.c  includes commands for
	  reloading, rotating logs, and status showing all threads
	- update_FFsource :  put files in src tree, also header files in /usr/include/

FFCLOCK  kernel config file now added to RADclock tarball and update_FFsource

Documentation
	- review and bringing up to date of all  {RAD,FF}clockDoc_* files
	- renaming extension  .c --> .txt
	- creation of RADclockDoc_guide.txt as a documentation, installation and
     userguide for both RADclock and FFclock
	- RADclockDoc_thread.c new sections:
		Networking and Sockets;   NTP_SERV Thread Detail
	- updating of README, AUTHORS, ChangeLog, NEWS
	- updating of RADclock manpages [ lots of changes ]
	- fixing and updating of RADclock copyright notices

+++ radclock/stampoutput.c
@@ -212,7 +212,8 @@ print_out_files(struct radclock_handle *handle, struct stamp_t *stamp)
 	/* Store generated stamp values */
 	if (handle->stampout_fd != NULL) {
 		err = fprintf(handle->stampout_fd,"%"VC_FMT" %.9Lf %.9Lf %"VC_FMT" %llu\n",
-				BST(stamp)->Ta, BST(stamp)->Tb, BST(stamp)->Te, BST(stamp)->Tf,
+				(long long unsigned)BST(stamp)->Ta, BST(stamp)->Tb, BST(stamp)->Te,
+				(long long unsigned)BST(stamp)->Tf,
 				(long long unsigned)stamp->id);
 		if (err < 0)
 			verbose(LOG_ERR, "Failed to write data to timestamp file");
@@ -229,15 +230,15 @@ print_out_files(struct radclock_handle *handle, struct stamp_t *stamp)
 		"%"VC_FMT" %"VC_FMT" %"VC_FMT" %.9lg %.9lg %.9lg %.11Lf "
 		"%.11Lf %.10lf %.10lf %.6lg %.6lg %.6lg %"VC_FMT" %u\n",
 		BST(stamp)->Tb,
-		BST(stamp)->Tf,
-		OUTPUT(handle, RTT),
+		(unsigned long long)BST(stamp)->Tf,
+		(unsigned long long)OUTPUT(handle, RTT),
 		OUTPUT(handle, phat),
 		OUTPUT(handle, plocal),
 		OUTPUT(handle, K),
 		OUTPUT(handle, thetahat),
-		OUTPUT(handle, RTThat),
-		OUTPUT(handle, RTThat_new),
-		OUTPUT(handle, RTThat_shift),
+		(unsigned long long)OUTPUT(handle, RTThat),
+		(unsigned long long)OUTPUT(handle, RTThat_new),
+		(unsigned long long)OUTPUT(handle, RTThat_shift),
 		OUTPUT(handle, th_naive),
 		OUTPUT(handle, minET),
 		OUTPUT(handle, minET_last),
